requirements.txt
APScheduler==3.11.0
boto3==1.40.29
fastapi==0.116.1
greenlet==3.2.4
llama-cloud-services==0.6.65
openai==1.100.0
pandas==2.3.1
pgvector==0.4.1
pip==25.1.1
psycopg2-binary==2.9.10
python-multipart==0.0.20
redis==6.4.0
sentence-transformers==5.1.0
uvicorn==0.35.0
